Send backend-specific ops in respective modules.
